refactor: refactors names list layout for reusability
